+++ tests/tests/pe_bliss2/address_converter_tests.cpp
@@ -63,7 +63,7 @@ TEST(AddressConverterTests, VaToRvaTest)
 {
 	address_converter ac(image_base);
 	EXPECT_EQ(ac.va_to_rva(0x456u), 0x456u - image_base);
-	EXPECT_EQ(ac.va_to_rva(0x456ull), 0x456ull - image_base);
+	EXPECT_EQ(ac.va_to_rva(static_cast<std::uint64_t>(0x456ull)), 0x456ull - image_base);
 	expect_throw_pe_error([&] {
 		(void)ac.va_to_rva(image_base - 1u);
 	}, address_converter_errc::address_conversion_overflow);
@@ -76,6 +76,6 @@ TEST(AddressConverterTests, VaToRvaTest2)
 {
 	address_converter ac(image_base);
 	expect_throw_pe_error([&] {
-		(void)ac.va_to_rva(0xff000000'00000000ull);
+		(void)ac.va_to_rva(static_cast<std::uint64_t>(0xff000000'00000000ull));
 	}, address_converter_errc::address_conversion_overflow);
 }
